requirements.txt
flask
gTTS
Pillow
mtranslate
